Fixes encoding issues :congratulations:
This fix will handle the encoding issues we've been having. The reason
for the errors turned out to be because the dagskammtur.is page is saved in
windows-1252 encoding and request doesn't handle it (and they aren't
sending the appropriate headers to notify anybody).

This patch will request the site in binary and convert that to
windows-1252 encoding before passing it on to the callback.